@@ -15,30 +15,25 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.2+3
+version: 1.0.3+4
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences: ^0.5.8
+  shared_preferences: ^2.0.5
   toast: ^0.1.5
-  playify:
-    git:
-      url: https://github.com/iberatkaya/playify.git
-      ref: playlists
-  font_awesome_flutter: ^8.8.1
-  flutter_redux: ^0.6.0
-  redux: ^4.0.0+3
-  flutter_statusbarcolor: ^0.2.3
-  pie_chart: ^3.1.1
-  permission_handler: ^5.0.1+1
-  app_settings: ^4.0.3
-  palette_generator: ^0.2.3
-  introduction_screen: ^1.0.9
-  flutter_xlider: ^3.4.0
+  playify: ^2.0.0
+  font_awesome_flutter: ^9.0.0
+  flutter_redux: ^0.8.2
+  redux: ^5.0.0
+  pie_chart: ^5.0.0
+  permission_handler: ^6.1.1
+  app_settings: ^4.1.0
+  palette_generator: ^0.3.0
+  introduction_screen: ^2.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
