@@ -3,5 +3,5 @@ import 'package:flutter/material.dart';
 class Action {
   dynamic type;
   dynamic payload;
-  Action({@required this.payload, @required this.type});
+  Action({required this.payload, required this.type});
 }
