@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_xlider/flutter_xlider.dart';
 import 'package:playify/playify.dart';
 import 'package:playify_app/utilities/utils.dart';
 
 class SliderEntry extends PopupMenuEntry {
   final double defaultVolume;
   SliderEntry({
-    @required this.defaultVolume,
-  }) : assert(defaultVolume != null);
+    required this.defaultVolume,
+  });
 
   @override
   State<StatefulWidget> createState() => SliderEntryState();
@@ -35,45 +34,25 @@ class SliderEntryState extends State<SliderEntry> {
   Widget build(BuildContext context) {
     return Container(
       height: 200,
-      width: 100,
-      child: FlutterSlider(
-        axis: Axis.vertical,
-        jump: true,
-        trackBar: FlutterSliderTrackBar(
-          inactiveTrackBar: BoxDecoration(
-            borderRadius: BorderRadius.circular(20),
-            color: Colors.blue[100],
-            border: Border.all(width: 3, color: Colors.blue),
-          ),
-          activeTrackBar: BoxDecoration(
-              borderRadius: BorderRadius.circular(4),
-              color: Theme.of(context).primaryColor),
+      width: 80,
+      child: RotatedBox(
+        quarterTurns: 3,
+        child: Slider(
+          value: volume,
+          min: 0,
+          max: 100,
+          onChanged: (newValue) async {
+            try {
+              final myVal = newValue / 100;
+              setState(() {
+                volume = newValue;
+              });
+              await playify.setVolume(myVal);
+            } catch (e) {
+              print(e);
+            }
+          },
         ),
-        step: FlutterSliderStep(step: 1),
-        values: [volume],
-        min: 0,
-        max: 100,
-        tooltip: FlutterSliderTooltip(
-            custom: (val) => Container(
-                color: themeModeColor(
-                    MediaQuery.of(context).platformBrightness, Colors.white),
-                padding: EdgeInsets.all(8),
-                child: Text(
-                  (val as double).toStringAsFixed(0) + "%",
-                )),
-            direction: FlutterSliderTooltipDirection.top),
-        rtl: true,
-        onDragging: (handlerIndex, lowerValue, upperValue) async {
-          try {
-            final myVal = lowerValue / 100;
-            setState(() {
-              volume = lowerValue;
-            });
-            await playify.setVolume(myVal);
-          } catch (e) {
-            print(e);
-          }
-        },
       ),
     );
   }
